+++ JoysOfEfficiency/Config.cs
@@ -44,5 +44,8 @@ namespace JoysOfEfficiency
         public int MachineRadius { get; set; } = 1;
         public bool AutoPetNearbyPets { get; set; } = false;
         public bool ProtectNectarProducingFlower { get; set; } = true;
+        public bool FPSCounter { get; set; } = true;
+        public int FPSCounterPosX { get; set; } = 16;
+        public int FPSCounterPosY { get; set; } = 192;
     }
 }