+++ JoysOfEfficiency/OptionsElements/ModifiedSlider.cs
@@ -73,7 +73,7 @@ namespace JoysOfEfficiency.OptionsElements
             label = $"{Label}: {Format(whichOption, Value + MinValue)}";
             greyedOut = IsDisabled();
 
-            base.draw(spriteBatch, slotX, slotY);
+            base.draw(spriteBatch, slotX, slotY - ((int)Game1.dialogueFont.MeasureString(label).Y - bounds.Height) / 2);
             IClickableMenu.drawTextureBox(spriteBatch, Game1.mouseCursors, OptionsSlider.sliderBGSource, slotX + bounds.X, slotY + bounds.Y, bounds.Width, bounds.Height, Color.White, Game1.pixelZoom, false);
             spriteBatch.Draw(Game1.mouseCursors, new Vector2(slotX + bounds.X + (bounds.Width - 10 * Game1.pixelZoom) * (Value / (float)MaxValue), slotY + bounds.Y), OptionsSlider.sliderButtonRect, Color.White, 0.0f, Vector2.Zero, Game1.pixelZoom, SpriteEffects.None, 0.9f);
         }